+++ sfall/FalloutEngine/VariableOffsets.h
@@ -258,6 +258,7 @@
 #define FO_VAR_procTableStrs              0x51C758  // table of procId (from define.h) => procName map
 #define FO_VAR_proto_main_msg_file        0x6647FC
 #define FO_VAR_proto_msg_files            0x6647AC
+#define FO_VAR_proto_none_str             0x664824
 #define FO_VAR_protoLists                 0x51C290
 #define FO_VAR_ptable                     0x59E934
 #define FO_VAR_pud                        0x59E960